gateway/requirements.txt
django==4.2.3 
cryptography==41.0.2
pymysql==1.1.0
python-decouple==3.8
asgiref==3.7.2
cffi==1.15.1
pycparser==2.21
sqlparse==0.4.4
typing_extensions==4.7.1
djangorestframework==3.14.0
pytz==2023.3
pika==1.3.2
requests==2.31.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
urllib3==2.0.4
